// frontbox-pin-console/web/src/types/generated/System.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { TraceRecord } from "./TraceRecord";

export type System = { id: number, name: string, active: boolean, 
/**
 * The most recent events this system emitted, oldest first
 */
recent_events: Array<TraceRecord>, };